Load all available forecasts submitted on forecast_dates from local hub repo.

This function will throw an error when no forecasts are submitted on any dates in forecast_dates for selected models, locations, types and target.

load_forecasts_repo(
  file_path,
  models = NULL,
  forecast_dates = NULL,
  locations = NULL,
  types = NULL,
  targets = NULL,
  hub = c("US", "ECDC"),
  verbose = TRUE
)

Arguments

file_path
path to the data-processed folder within a local clone of the hub repo

models
Character vector of model abbreviations. Default all models that submitted forecasts meeting the other criteria.

forecast_dates
The forecast date of forecasts to retrieve. Default to all valid forecast dates. The function will throw an error if all dates in this parameter are invalid forecast dates.

locations
list of fips. Default to all locations with available forecasts.

types
Character vector specifying type of forecasts to load: "quantile" and/or "point". Default to all valid forecast types.

targets
character vector of targets to retrieve, for example c('1 wk ahead cum death', '2 wk ahead cum death'). Default to NULL which stands for all valid targets.

hub
character vector, where the first element indicates the hub from which to load forecasts. Possible options are "US" and "ECDC".

verbose
logical to print out diagnostic messages. Default is TRUE

Value

data.frame with columns model, forecast_date, location, horizon, temporal_resolution, target_variable, target_end_date, type, quantile, value, location_name, population, geo_type, geo_value, abbreviation

Details

This function will drop rows with NULL in value column.