Load the most recent forecasts of all that were submitted on forecast_dates from a local clone of reichlab/covid19-forecast-hub repo.

This function will drop rows with NULLs in value column.

load_latest_forecasts_repo(
  file_path,
  models = NULL,
  forecast_dates,
  locations = NULL,
  types = NULL,
  targets = NULL,
  hub = c("US", "ECDC"),
  verbose = TRUE
)

Arguments

file_path
path to the data-processed folder within a local clone of the hub repo

models
Character vector of model abbreviations. Default all models that submitted forecasts meeting the other criteria.

forecast_dates
A 2 dimensional list of forecast dates to retrieve forecasts. This function will return the latest forecasts for each sub-list of dates. Default to NULL which would include all valid forecast dates. The function will throw an error if all dates in this parameter are invalid forecast dates.

locations
list of fips. Default to all locations with available forecasts.

types
Character vector specifying type of forecasts to load: "quantile" and/or "point". Default to all valid forecast types.

targets
character vector of targets to retrieve, for example c('1 wk ahead cum death', '2 wk ahead cum death'). Default to NULL which stands for all valid targets.

hub
character vector, where the first element indicates the hub from which to load forecasts. Possible options are "US" and "ECDC".

verbose
logical to print out diagnostic messages. Default is TRUE

Details

Please use load_forecasts_repo() instead.