Load truth data under multiple target variables from multiple truth sources using files in reichlab/covid19-forecast-hub.

"inc hosp" is only available from "HeatlthData" and this function is not loading data for other target variables from "HealthData".

load_truth(
  truth_source = NULL,
  target_variable = NULL,
  as_of = NULL,
  truth_end_date = NULL,
  temporal_resolution = NULL,
  locations = NULL,
  data_location = NULL,
  local_repo_path = NULL,
  hub = c("US", "ECDC")
)

Arguments

truth_source
character vector specifying where the truths will be loaded from: currently support "JHU", "USAFacts", "NYTimes", "HealthData" and "ECDC". If NULL, default for US hub is c("JHU", "HealthData"). If NULL, default for ECDC hub is c("JHU").

target_variable
string specifying target type It should be one or more of "cum death", "inc case", "inc death", "inc hosp". If NULL, default for US hub is c("inc case", "inc death", "inc hosp"). If NULL, default for ECDC hub is c("inc case", "inc death").

as_of
character vector of "as of" dates to use for querying truths in format 'yyyy-mm-dd'. For each spatial unit and temporal reporting unit, the last available data with an issue date on or before the given as_of date are returned. This is only available for covidData now.

truth_end_date
date to include the last available truth point in 'yyyy-mm-dd' format. If NULL,default to system date.

temporal_resolution
character specifying temporal resolution to include: currently support "weekly" and "daily". If NULL, default to "weekly" for cases and deaths, "daily" for hospitalizations. Weekly temporal_resolution will not be applied to "inc hosp" when multiple target variables are specified. "ECDC" truth data is weekly by default. Daily level data is not available.

locations
vector of valid location code. If NULL, default to all locations with available forecasts. US hub is using FIPS code and ECDC hub is using country name abbreviation.

data_location
character specifying the location of truth data. Currently only supports "local_hub_repo", "remote_hub_repo" and "covidData". If NULL, default to "remote_hub_repo".

local_repo_path
path to local clone of the reichlab/covid19-forecast-hub repository. Only used when data_location is "local_hub_repo"

hub
character, which hub to use. Default is "US", other option is "ECDC"

Value

data.frame with columns model, inc_cum, death_case, target_end_date, location, value, location_name, population, geo_type, geo_value, abbreviation

Details

When loading data for multiple target_variables, temporal_resolution will be applied to all target variables but "inc hosp". In that case, this function will return daily incident hospitalization counts along with other data.

Weekly temporal resolution will be applied to "inc hosp" if the user specifies "inc hosp" as the only target_variable.

When loading weekly data, if there are not enough observations for a week, the corresponding weekly count would be NA in resulting data frame.

Examples

library(covidHubUtils)

# load for US
load_truth(
  truth_source = c("JHU", "HealthData"),
  target_variable = c("inc case", "inc death", "inc hosp")
)
#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   date = col_date(format = ""),
#>   location = col_character(),
#>   location_name = col_character(),
#>   value = col_double()
#> )
#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   date = col_date(format = ""),
#>   location = col_character(),
#>   location_name = col_character(),
#>   value = col_double()
#> )
#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   date = col_date(format = ""),
#>   location = col_character(),
#>   location_name = col_character(),
#>   value = col_double()
#> )
#> # A tibble: 512,564 x 11
#>    model target_variable target_end_date location value location_name population
#>    <chr> <chr>           <date>          <chr>    <dbl> <chr>              <dbl>
#>  1 Obse… inc hosp        2020-07-21      19          33 Iowa             3155070
#>  2 Obse… inc hosp        2020-07-21      72           0 Puerto Rico      3754939
#>  3 Obse… inc hosp        2020-07-19      01         136 Alabama          4903185
#>  4 Obse… inc hosp        2020-07-17      19          20 Iowa             3155070
#>  5 Obse… inc hosp        2020-07-17      53          26 Washington       7614893
#>  6 Obse… inc hosp        2020-07-16      41          11 Oregon           4217737
#>  7 Obse… inc hosp        2020-07-15      15           2 Hawaii           1415872
#>  8 Obse… inc hosp        2021-06-13      78           0 Virgin Islan…     107268
#>  9 Obse… inc hosp        2020-07-25      09           8 Connecticut      3565287
#> 10 Obse… inc hosp        2020-08-23      28          82 Mississippi      2976149
#> # … with 512,554 more rows, and 4 more variables: geo_type <chr>,
#> #   geo_value <chr>, abbreviation <chr>, full_location_name <chr>

# load for ECDC
load_truth(
  truth_source = c("JHU"),
  target_variable = c("inc case", "inc death"),
  hub = "ECDC"
)
#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   location = col_character(),
#>   location_name = col_character(),
#>   date = col_date(format = ""),
#>   value = col_double()
#> )
#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   location = col_character(),
#>   location_name = col_character(),
#>   date = col_date(format = ""),
#>   value = col_double()
#> )
#> # A tibble: 4,928 x 7
#>    model target_variable target_end_date location value location_name population
#>    <chr> <chr>           <date>          <chr>    <dbl> <chr>              <int>
#>  1 Obse… inc case        2020-01-25      AT           0 Austria          8809212
#>  2 Obse… inc case        2020-02-01      AT           0 Austria          8809212
#>  3 Obse… inc case        2020-02-08      AT           0 Austria          8809212
#>  4 Obse… inc case        2020-02-15      AT           0 Austria          8809212
#>  5 Obse… inc case        2020-02-22      AT           0 Austria          8809212
#>  6 Obse… inc case        2020-02-29      AT           9 Austria          8809212
#>  7 Obse… inc case        2020-03-07      AT          70 Austria          8809212
#>  8 Obse… inc case        2020-03-14      AT         576 Austria          8809212
#>  9 Obse… inc case        2020-03-21      AT        2159 Austria          8809212
#> 10 Obse… inc case        2020-03-28      AT        5457 Austria          8809212
#> # … with 4,918 more rows