Plot forecasts and optional truth for only one selected target variable. Faceted plots for multiple models, locations and forecast dates are supported with specified facet formula.

Plot forecasts and optional truth for only one selected target variable. Faceted plots for multiple models, locations and forecast dates are supported with specified facet formula.

plot_forecasts(
  forecast_data,
  truth_data = NULL,
  hub = c("US", "ECDC"),
  models,
  target_variable,
  locations,
  facet = NULL,
  facet_scales = "fixed",
  facet_nrow = NULL,
  facet_ncol = NULL,
  forecast_dates,
  intervals,
  horizon,
  truth_source,
  use_median_as_point = FALSE,
  plot_truth = TRUE,
  plot = TRUE,
  fill_by_model = FALSE,
  fill_transparency = 1,
  truth_as_of = NULL,
  title = "default",
  subtitle = "default",
  show_caption = TRUE
)

Arguments

forecast_data
data frame with truth and forecasts from load_forecasts()

truth_data
optional data frame from one truth source in the format returned by load_truth(). It needs to have columns model, target_variable, target_end_date, location and value. Model column can be "Observed Data (a truth source)".

hub
character, which hub to use. Default is "US", other option is "ECDC".

models
vector of strings specifying models to plot. Default to all models in forecast_data.

target_variable
string specifying target type. It should be one of "cum death", "inc case", "inc death" and "inc hosp". "cum death" and "inc hosp" are only available in forecasts from US hub now. If there is only one target_variable in forecast_data, this parameter is optional.

locations
string for fips code, 'US' or abbreviation of European countries. Default to all locations in forecast_data.

facet
interpretable facet option for ggplot. Function will error if multiple locations are passed in without location in the facet formula.

facet_scales
argument for scales in ggplot2::facet_wrap. Default to "fixed".

facet_nrow
number of rows for facetting; optional.

facet_ncol
number of columns for facetting; optional.

forecast_dates
date string vectors for forecast dates to plot. Default to forecast_dates present in the data.

intervals
values indicating which central prediction interval levels to plot. NULL means only plotting point forecasts. If not provided, it will default to c(.5, .8, .95). When plotting 6 models or more, the plot will be reduced to show .95 interval only.

horizon
forecasts are plotted for the horizon time steps after the forecast date. Default to all available horizons in forecast data.

truth_source
character specifying where the truth data will be loaded from if truth_data is not provided. Currently support "JHU", "USAFacts", "NYTimes", "HealthData" and "ECDC". Optional if truth_data is provided.

use_median_as_point
boolean for using median quantiles as point forecasts in plot. Default to FALSE.

plot_truth
boolean for showing truth data in plot. Default to FALSE.

plot
boolean for showing the plot. Default to TRUE. Currently supports "JHU","USAFacts", "NYTimes". Default to "JHU".

fill_by_model
boolean for specifying colors in plot. If TRUE, separate colors will be used for each model. If FALSE, only blues will be used for all models. Default to FALSE.

fill_transparency
numeric value used to set transparency of intervals. 0 means fully transparent, 1 means opaque.

truth_as_of
the plot includes the truth data that would have been in real time as of the truth_as_of date (not using this parameter when truth data is from github repo)

title
optional text for the title of the plot. If left as "default", the title will be automatically generated. If "none", no title will be plotted.

subtitle
optional text for the subtitle of the plot. If left as "default", the subtitle will be automatically generated. If "none", no subtitle will be plotted.

show_caption
logical, if TRUE, caption will be included showing data sources

Value

invisible ggplot object

Examples

if (FALSE) {
us_forecasts <-load_latest_forecasts(
  models =c("COVIDhub-ensemble","COVIDhub-baseline",
  "JHU_CSSE-DECOM","CovidAnalytics-DELPHI",
  "LANL-GrowthRate", "Microsoft-DeepSTIA"),
  last_forecast_date = Sys.Date(),
  forecast_date_window_size = 6,
  locations = "US",
  types = c("point","quantile"),
  targets = paste(1:4, "wk ahead inc case"),
  source = "zoltar")
 
plot_forecasts(forecast_data = us_forecasts,
  models = c("COVIDhub-ensemble","COVIDhub-baseline","JHU_CSSE-DECOM",
  "CovidAnalytics-DELPHI","LANL-GrowthRate", "Microsoft-DeepSTIA"),
  target_variable = "inc case", locations = c("US"), intervals = c(.95),
  horizon = 4, truth_source = "JHU", plot = TRUE, fill_by_model = TRUE,
  facet = ~model, facet_scales = "fixed", title = "default",
  show_caption = TRUE)

ecdc_forecasts <- load_latest_forecasts(models=c("ILM-EKF"),
  hub = c("ECDC","US"), last_forecast_date = "2021-03-08",
  forecast_date_window_size = 0,
  locations = c("GB"),
  targets = paste(1:4, "wk ahead inc death"),
  source = "zoltar")
plot_forecasts(forecast_data = ecdc_forecasts, 
               hub  = c("ECDC","US"), 
               locations ="GB", 
               truth_source = "JHU")
}