Reformat forecast data frame returned from zoltar query This function will throw a warning and return an empty data.frame with columns model, forecast_date, location, type, quantile and value, when no forecasts are available in zoltar_query_result.

Reformat forecast data frame returned from zoltar query This function will throw a warning and return an empty data.frame with columns model, forecast_date, location, type, quantile and value, when no forecasts are available in zoltar_query_result.

reformat_forecasts(zoltar_query_result, verbose = TRUE)

Arguments

zoltar_query_result
dataframe returned by zoltr::do_zoltar_query

verbose
logical for printing error messages when checking zoltar_query_result. Default to TRUE.

Value

data.frame with columns model, forecast_date, location, horizon, temporal_resolution, target_variable, target_end_date, type, quantile, value