Load all available forecasts submitted on forecast_dates from Zoltar or local hub repo.

Load all available forecasts submitted on forecast_dates from Zoltar or local hub repo.

load_forecasts(
  models = NULL,
  forecast_dates = NULL,
  locations = NULL,
  types = NULL,
  targets = NULL,
  source = "zoltar",
  hub_repo_path,
  data_processed_subpath = "data-processed/",
  as_of = NULL,
  hub = c("US", "ECDC"),
  verbose = TRUE
)

Arguments

models
Character vector of model abbreviations. Default all models that submitted forecasts meeting the other criteria.

forecast_dates
The forecast date of forecasts to retrieve. Default to all valid forecast dates. The function will throw an error if all dates in this parameter are invalid forecast dates in Zoltar.

locations
list of location codes. Default to all locations with available forecasts.

types
Character vector specifying type of forecasts to load: "quantile" and/or "point". Default to all valid forecast types.

targets
character vector of targets to retrieve, for example c('1 wk ahead cum death', '2 wk ahead cum death'). Default to NULL which stands for all valid targets.

source
string specifying where forecasts will be loaded from: either "local_hub_repo" or "zoltar". Default to "zoltar".

hub_repo_path
path to local clone of the forecast hub repository

data_processed_subpath
folder within the hub_repo_path that contains forecast submission files. Default to "data-processed/", which is appropriate for the covid19-forecast-hub repository.

as_of
character for date time to load forecasts submitted as of this time from Zoltar. Ignored if source = "local_hub_repo". It could use the format of one of the three examples: "2021-01-01", "2020-01-01 01:01:01" and "2020-01-01 01:01:01 UTC". If you would like to set a timezone, it has to be UTC now. If not, helper function will append the default timezone to your input based on hub parameter. Default to NULL to load the latest version.

hub
character vector, where the first element indicates the hub from which to load forecasts. Possible options are "US" and "ECDC".

verbose
logical to print out diagnostic messages. Default is TRUE.

Value

data.frame with columns model, forecast_date, location, horizon, temporal_resolution, target_variable, target_end_date, type, quantile, value, location_name, population, geo_type, geo_value, abbreviation

Examples

# load forecasts from US forecast hub
load_forecasts(models = "COVIDhub-ensemble",
 forecast_date = "2020-12-07",
 locations = "US",
 types = c("point","quantile"),
 targets = paste(1:4, "wk ahead inc case"),
 source = "zoltar",
 verbose = FALSE,
 as_of = NULL)
#> get_token(): POST: https://zoltardata.com/api-token-auth/
#> get_resource(): GET: https://zoltardata.com/api/projects/
#> get_resource(): GET: https://zoltardata.com/api/project/44/timezeros/
#> get_resource(): GET: https://zoltardata.com/api/job/73263/
#> get_resource(): GET: https://zoltardata.com/api/job/73263/
#> get_resource(): GET: https://zoltardata.com/api/job/73263/data/
#> # A tibble: 32 x 16
#>    model        forecast_date location horizon temporal_resolut… target_variable
#>    <chr>        <date>        <chr>    <chr>   <chr>             <chr>          
#>  1 COVIDhub-en… 2020-12-07    US       1       wk                inc case       
#>  2 COVIDhub-en… 2020-12-07    US       1       wk                inc case       
#>  3 COVIDhub-en… 2020-12-07    US       1       wk                inc case       
#>  4 COVIDhub-en… 2020-12-07    US       1       wk                inc case       
#>  5 COVIDhub-en… 2020-12-07    US       1       wk                inc case       
#>  6 COVIDhub-en… 2020-12-07    US       1       wk                inc case       
#>  7 COVIDhub-en… 2020-12-07    US       1       wk                inc case       
#>  8 COVIDhub-en… 2020-12-07    US       1       wk                inc case       
#>  9 COVIDhub-en… 2020-12-07    US       2       wk                inc case       
#> 10 COVIDhub-en… 2020-12-07    US       2       wk                inc case       
#> # … with 22 more rows, and 10 more variables: target_end_date <date>,
#> #   type <chr>, quantile <dbl>, value <dbl>, location_name <chr>,
#> #   population <dbl>, geo_type <chr>, geo_value <chr>, abbreviation <chr>,
#> #   full_location_name <chr>

# load forecasts from ECDC forecast hub
load_forecasts(models = "ILM-EKF",
 hub = c("ECDC","US"),
 forecast_date = "2021-03-08",
 locations = "GB",
 targets = paste(1:4, "wk ahead inc death"),
 source = "zoltar")
#> get_token(): POST: https://zoltardata.com/api-token-auth/
#> get_resource(): GET: https://zoltardata.com/api/projects/
#> get_resource(): GET: https://zoltardata.com/api/project/238/timezeros/
#> polling for status change. job_url=https://zoltardata.com/api/job/73264/
#> get_resource(): GET: https://zoltardata.com/api/job/73264/
#> QUEUED
#> get_resource(): GET: https://zoltardata.com/api/job/73264/
#> SUCCESS
#> get_resource(): GET: https://zoltardata.com/api/job/73264/data/
#> # A tibble: 96 x 12
#>    model   forecast_date location horizon temporal_resolution target_variable
#>    <chr>   <date>        <chr>    <chr>   <chr>               <chr>          
#>  1 ILM-EKF 2021-03-08    GB       1       wk                  inc death      
#>  2 ILM-EKF 2021-03-08    GB       1       wk                  inc death      
#>  3 ILM-EKF 2021-03-08    GB       1       wk                  inc death      
#>  4 ILM-EKF 2021-03-08    GB       1       wk                  inc death      
#>  5 ILM-EKF 2021-03-08    GB       1       wk                  inc death      
#>  6 ILM-EKF 2021-03-08    GB       1       wk                  inc death      
#>  7 ILM-EKF 2021-03-08    GB       1       wk                  inc death      
#>  8 ILM-EKF 2021-03-08    GB       1       wk                  inc death      
#>  9 ILM-EKF 2021-03-08    GB       1       wk                  inc death      
#> 10 ILM-EKF 2021-03-08    GB       1       wk                  inc death      
#> # … with 86 more rows, and 6 more variables: target_end_date <date>,
#> #   type <chr>, quantile <dbl>, value <dbl>, location_name <chr>,
#> #   population <int>